requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-jose 
cryptography
httpx 
supabase
langgraph
langchain-core
langchain-openai
